+++ quri_algo/algo/phase_estimation/spe/utils/fourier/gaussian_coefficient.py
@@ -18,6 +18,27 @@ import numpy.typing as npt
 from .coefficient import FouierCoefficientSampler, SPEFourierCoefficient
 
 
+@dataclass
+class GaussianParam:
+    r"""Hyperparameters of the Gaussian distribution function.
+
+    .. math::
+        G_{\sigma}(x) = \int_{-T}^{T} e^{-\frac{1}{2}(\sigma^2 \pi^2 k^2)} e^{2\pi i k x} dk
+
+    Args:
+        T: The boundary of the Fourier integration region.
+        N: Number of points to disceretize the integration interval.
+        sigma: Standard deviation of the Gaussian distribution.
+        n_sample: Number of samples to sample the Fourier coefficient
+            :math:`e^{-\frac{1}{2}(\sigma^2 \pi^2 k^2)}`
+    """
+
+    T: float
+    N: int
+    sigma: float
+    n_sample: int
+
+
 def get_kn(T: float, N: int) -> npt.NDArray[np.float64]:
     n = np.arange(N)
     return -T + 2 * T / N * n
